[Fixes #1841] Change XML DCS and XmlSerializer output formatters to not derive from the base XmlOutputFormatter
