+++ src/Microsoft.AspNet.Mvc.Core/Formatters/XmlSerializerOutputFormatter.cs
@@ -2,9 +2,11 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.IO;
 using System.Threading.Tasks;
 using System.Xml;
 using System.Xml.Serialization;
+using Microsoft.Net.Http.Headers;
 
 namespace Microsoft.AspNet.Mvc
 {
@@ -12,14 +14,14 @@ namespace Microsoft.AspNet.Mvc
     /// This class handles serialization of objects
     /// to XML using <see cref="XmlSerializer"/>
     /// </summary>
-    public class XmlSerializerOutputFormatter : XmlOutputFormatter
+    public class XmlSerializerOutputFormatter : OutputFormatter
     {
         /// <summary>
         /// Initializes a new instance of <see cref="XmlSerializerOutputFormatter"/>
         /// with default XmlWriterSettings.
         /// </summary>
-        public XmlSerializerOutputFormatter()
-            : this(GetDefaultXmlWriterSettings())
+        public XmlSerializerOutputFormatter() :
+            this(FormattingUtilities.GetDefaultXmlWriterSettings())
         {
         }
 
@@ -28,8 +30,39 @@ namespace Microsoft.AspNet.Mvc
         /// </summary>
         /// <param name="writerSettings">The settings to be used by the <see cref="XmlSerializer"/>.</param>
         public XmlSerializerOutputFormatter([NotNull] XmlWriterSettings writerSettings)
-            : base(writerSettings)
         {
+            SupportedEncodings.Add(Encodings.UTF8EncodingWithoutBOM);
+            SupportedEncodings.Add(Encodings.UTF16EncodingLittleEndian);
+
+            SupportedMediaTypes.Add(MediaTypeHeaderValue.Parse("application/xml"));
+            SupportedMediaTypes.Add(MediaTypeHeaderValue.Parse("text/xml"));
+
+            WriterSettings = writerSettings;
+        }
+
+        /// <summary>
+        /// Gets the settings to be used by the XmlWriter.
+        /// </summary>
+        public XmlWriterSettings WriterSettings { get; }
+
+        /// <summary>
+        /// Gets the type of the object to be serialized.
+        /// </summary>
+        /// <param name="declaredType">The declared type.</param>
+        /// <param name="runtimeType">The runtime type.</param>
+        /// <returns>The type of the object to be serialized.</returns>
+        protected virtual Type GetSerializableType(Type declaredType, Type runtimeType)
+        {
+            if (declaredType == null ||
+                declaredType == typeof(object))
+            {
+                if (runtimeType != null)
+                {
+                    return runtimeType;
+                }
+            }
+
+            return declaredType;
         }
 
         /// <inheritdoc />
@@ -56,6 +89,17 @@ namespace Microsoft.AspNet.Mvc
                 // false, then this Formatter is not picked up at all.
                 return null;
             }
+        }
+
+        /// <summary>
+        /// Creates a new instance of <see cref="XmlWriter"/> using the given stream and the WriterSettings.
+        /// </summary>
+        /// <param name="writeStream">The stream on which the XmlWriter should operate on.</param>
+        /// <returns>A new instance of <see cref="XmlWriter"/></returns>
+        public virtual XmlWriter CreateXmlWriter([NotNull] Stream writeStream,
+                                                 [NotNull] XmlWriterSettings xmlWriterSettings)
+        {
+            return XmlWriter.Create(writeStream, xmlWriterSettings);
         }
 
         /// <inheritdoc />
