+++ test/Microsoft.AspNet.Mvc.Core.Test/Formatters/XmlSerializerOutputFormatterTests.cs
@@ -72,7 +72,7 @@ namespace Microsoft.AspNet.Mvc.Core
             // Arrange
             var sampleInput = new DummyClass { SampleInt = 10 };
             var formatter = new XmlSerializerOutputFormatter(
-                XmlOutputFormatter.GetDefaultXmlWriterSettings());
+                FormattingUtilities.GetDefaultXmlWriterSettings());
             var outputFormatterContext = GetOutputFormatterContext(sampleInput, sampleInput.GetType());
 
             // Act
@@ -102,7 +102,7 @@ namespace Microsoft.AspNet.Mvc.Core
                 }
             };
             var formatter = new XmlSerializerOutputFormatter(
-                XmlOutputFormatter.GetDefaultXmlWriterSettings());
+                FormattingUtilities.GetDefaultXmlWriterSettings());
             var outputFormatterContext = GetOutputFormatterContext(sampleInput, sampleInput.GetType());
 
             // Act
@@ -153,7 +153,7 @@ namespace Microsoft.AspNet.Mvc.Core
             var outputFormatterContext =
                 GetOutputFormatterContext(sampleInput, sampleInput.GetType(), "application/xml; charset=utf-16");
             var formatter = new XmlSerializerOutputFormatter(
-                XmlOutputFormatter.GetDefaultXmlWriterSettings());
+                FormattingUtilities.GetDefaultXmlWriterSettings());
             formatter.WriterSettings.OmitXmlDeclaration = false;
 
             // Act
@@ -175,7 +175,7 @@ namespace Microsoft.AspNet.Mvc.Core
             // Arrange
             var sampleInput = new DummyClass { SampleInt = 10 };
             var formatter = new XmlSerializerOutputFormatter(
-                XmlOutputFormatter.GetDefaultXmlWriterSettings());
+                FormattingUtilities.GetDefaultXmlWriterSettings());
             formatter.WriterSettings.Indent = true;
             var outputFormatterContext = GetOutputFormatterContext(sampleInput, sampleInput.GetType());
 
@@ -198,7 +198,7 @@ namespace Microsoft.AspNet.Mvc.Core
             // Arrange
             var sampleInput = new DummyClass { SampleInt = 10 };
             var formatter = new XmlSerializerOutputFormatter(
-                XmlOutputFormatter.GetDefaultXmlWriterSettings());
+                FormattingUtilities.GetDefaultXmlWriterSettings());
             var outputFormatterContext = GetOutputFormatterContext(sampleInput, sampleInput.GetType());
 
             // Act
